Fix Grafana dashboard UTC time display and database schema
- Fixed transceivers frequency column from INTEGER to BIGINT to handle large VATSIM frequency values
- Updated Grafana timezone configuration to UTC for proper time display
- Modified dashboard SQL queries to display actual UTC timestamps from database
- Added UTC labels to dashboard panel titles for clarity
- Created migration file 012_fix_transceivers_frequency_column.sql
- Added explanatory comments to Grafana configuration files

+++ database/init.sql
@@ -232,7 +232,7 @@ CREATE TABLE IF NOT EXISTS transceivers (
     id SERIAL PRIMARY KEY,
     callsign VARCHAR(50) NOT NULL,
     transceiver_id INTEGER NOT NULL,
-    frequency INTEGER NOT NULL,
+    frequency BIGINT NOT NULL,
     position_lat DOUBLE PRECISION,
     position_lon DOUBLE PRECISION,
     height_msl DOUBLE PRECISION,
